+++ src/SegundoObligatorio/JSONWeatherMappable.swift
@@ -5,7 +5,7 @@ class JSONWeatherMappable: Mappable {
     var name:String=""
     var iconId:Int=0
     var iconString:String=""
-    var temperature:Double=0.0
+    var temperature:Int=0
     required init? (map: Map){
     }
